Fixed wall rendering glitch + added props
Biggest change was updating the wall rendering so they know longer jut out at you when walking along side them, unless in a 1x1 hallway in which case I think it is good for the aesthetic as it appears more cramped. I also added "props" which are render differently then walls and are essentially tables and such non interactive items.